Добавлены поля StarTime и Duration ÐРеализация ТЗ 8 спринатады работы с ними в TaskManager, расширены тесты
