+++ src/exception/FileManagerCrossedTimeInTasksException.java
@@ -1,0 +1,7 @@
+package exception;
+
+public class FileManagerCrossedTimeInTasksException extends RuntimeException {
+    public FileManagerCrossedTimeInTasksException(String message) {
+        super(message);
+    }
+}
