+++ src/manager/TaskManager.java
@@ -46,4 +46,6 @@ public interface TaskManager {
     void updateEpic(Epic epic);
 
     List<Task> getHistory();
+
+    List<Task> getPrioritizedTasks();
 }
