+++ src/tasks/Epic.java
@@ -1,10 +1,16 @@
 package tasks;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 
 public class Epic extends Task {
 
     private ArrayList<Integer> subTaskIds;
+    private LocalDateTime endTime;
+
+    public void setEndTime(LocalDateTime endTime) {
+        this.endTime = endTime;
+    }
 
     public Epic(String name, String description) {
         super(name, description, Status.NEW);
@@ -26,5 +32,10 @@ public class Epic extends Task {
     public ArrayList<Integer> getSubTaskIds() {
         ArrayList<Integer> returnedSubTaskIds = subTaskIds;
         return returnedSubTaskIds;
+    }
+
+    @Override
+    public LocalDateTime getEndTime() {
+        return endTime;
     }
 }
